feat: Cambiar importaciones

+++ app/src/main/java/com/example/mimascota/Model/Accesorios.kt
@@ -4,7 +4,7 @@ data class Accesorios(
     val idPro: String,
     val namePro: String,
     val descriptionPro: String,
-    val pricePro: Double,
+    val pricePro: Int,
     val tipoAccesorio: String,
     val material: String
 ) : Product(idPro, namePro, descriptionPro, pricePro) {
